real vps/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
docker==7.1.0
aiohttp==3.9.5
